linfe: Acertos para sincronização de endereços

+++ database/migrations/2021_04_03_020325_create_produtos_table.php
@@ -31,7 +31,7 @@ class CreateProdutosTable extends Migration
             $table->decimal('peso')->nullable();
             $table->string('ncm');
             $table->string('sku');
-            $table->integer('li_id')->nullable();
+            $table->bigInteger('li_id')->nullable();
 
             $table->timestamps();
         });
